actualizado readme

+++ aux_scripts/usbdevgone.sh
@@ -25,13 +25,9 @@ then
 	done
 fi
 
-#if file don't exists
-if ! test -e $filepath
+#if file don't exists and polkit version is < 0.106, remove pkla file to disable polkit rule
+if ! test -e $filepath && test $(pkaction --version | cut -d " " -f 3 | cut -d "." -f 2) -lt 106
 then
-	#if polkit version is < 0.106, remove pkla file to disable polkit rule
-	if test $(pkaction --version | cut -d " " -f 3 | cut -d "." -f 2) -lt 106
-	then
-		rm /etc/polkit-1/localauthority/50-local.d/50-inhibit-shutdown.pkla
-		service polkit restart
-	fi	
+	rm /etc/polkit-1/localauthority/50-local.d/50-inhibit-shutdown.pkla
+	service polkit restart	
 fi
